+++ common/pixel.h
@@ -53,7 +53,7 @@ static const struct {
     {  4,  4 }
 };
 
-static const int x264_size2pixel[5][5] = {
+static const uint8_t x264_size2pixel[5][5] = {
     { 0, },
     { 0, PIXEL_4x4, PIXEL_8x4, 0, 0 },
     { 0, PIXEL_4x8, PIXEL_8x8, 0, PIXEL_16x8 },
